+++ satella/coding/decorators/__init__.py
@@ -2,7 +2,7 @@ from .arguments import auto_adapt_to_methods, attach_arguments, for_argument, \
     execute_before, copy_arguments, replace_argument_if, transform_result, \
     transform_arguments
 from .decorators import wraps, chain_functions, has_keys, short_none, memoize, return_as_list, \
-    default_return, cache_memoize
+    default_return, cache_memoize, call_method_on_exception
 from .flow_control import loop_while, queue_get
 from .preconditions import postcondition, precondition
 from .retry_dec import retry
@@ -12,4 +12,4 @@ __all__ = ['retry', 'transform_result', 'transform_arguments',
            'chain_functions', 'has_keys', 'short_none', 'auto_adapt_to_methods',
            'attach_arguments', 'for_argument', 'loop_while', 'memoize',
            'copy_arguments', 'replace_argument_if', 'return_as_list',
-           'default_return', 'cache_memoize']
+           'default_return', 'cache_memoize', 'call_method_on_exception']