+++ tests/test_coding/test_monitor.py
@@ -156,3 +156,32 @@ class MonitorTest(unittest.TestCase):
         while a.is_alive() or b.is_alive():
             sleep(0.1)
             self.assertNotEqual(q.qsize(), 2)
+
+    def test_monitoring_synchronize_on_attribute(self):
+        class TestedClass:
+            def __init__(self, cqueue):
+                self.cqueue = cqueue
+                self.monitor = Monitor()
+
+            @Monitor.synchronize_on_attribute('monitor')
+            def execute(self):
+                self.cqueue.put(1)
+                sleep(1)
+                self.cqueue.get()
+
+        class TesterThread(Thread):
+            def __init__(self, tc):
+                self.tc = tc
+                Thread.__init__(self)
+
+            def run(self):
+                self.tc.execute()
+
+        q = Queue()
+        tc = TestedClass(q)
+        a, b = TesterThread(tc), TesterThread(tc)
+        a.start(), b.start()
+
+        while a.is_alive() or b.is_alive():
+            sleep(0.1)
+            self.assertNotEqual(q.qsize(), 2)