+++ satella/coding/concurrent/futures/__init__.py
@@ -1,6 +1,7 @@
 from .call_in_future import call_in_future
 from .futures import Future, WrappingFuture, InvalidStateError, wrap_if
 from .wrapped_executor import ExecutorWrapper
+from .collection import FutureCollection
 
 __all__ = ['Future', 'WrappingFuture', 'InvalidStateError', 'call_in_future',
-           'ExecutorWrapper', 'wrap_if']
+           'ExecutorWrapper', 'wrap_if', 'FutureCollection']