+++ satella/coding/concurrent/queue.py
@@ -7,7 +7,7 @@ from satella.coding.recast_exceptions import rethrow_as
 from satella.coding.concurrent.thread import Condition
 from satella.coding.typing import T
 from satella.exceptions import WouldWaitMore, Empty
-from satella.time import measure
+from satella.time.measure import measure
 
 
 class PeekableQueue(tp.Generic[T]):