+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v2.17.5
 
 * added terminate_on application to prepare in TerminableThread
+* added call_method_on_exception