+++ CHANGELOG.md
@@ -1,1 +1,4 @@
 # v2.17.6
+
+* added `synchronize_on_attribute` for `Monitor`
+