+++ CHANGELOG.md
@@ -1,1 +1,4 @@
 # v2.17.7
+
+* added new things to `ExponentialBackoff`
+* fixed longer first waits in `ExponentialBackoff`