+++ CHANGELOG.md
@@ -1,1 +1,4 @@
 # v2.17.8
+
+* bugfix release for 2.17.8
+    * some modules were unimportable